+++ Stitch/Graph/StitchAI/Mapping/SyntaxToActions/deriveActions.swift
@@ -9,156 +9,49 @@ import Foundation
 import SwiftUI
 
 
-// ──────────────────────────────────────────────────────────────
-// MARK: - Mapping logic
-// ──────────────────────────────────────────────────────────────
 extension SyntaxView {
-    
-    /// Leaf-level mapping for **this** node only
-    private func deriveLayer() -> (layer: VPLLayer,
-                                   extras: VPLLayerConcepts) {
 
-        // ── Base mapping from SyntaxViewName → Layer ────────────────────────
-        var layerType: Layer
-        var extras: VPLLayerConcepts = []
-
-        switch self.name {
-        case .rectangle:         layerType = .rectangle
-        case .circle:            layerType = .oval
-        case .text:              layerType = .text
-        case .image:             layerType = .image
-
-        case .hStack:
-            layerType = .group
-            extras.append(
-                .layerInputSet(VPLLayerInputSet(id: id,
-                                                input: .orientation,
-                                                value: .orientation(.horizontal)))
-            )
-
-        case .vStack:
-            layerType = .group
-            extras.append(
-                .layerInputSet(VPLLayerInputSet(id: id,
-                                                input: .orientation,
-                                                value: .orientation(.vertical)))
-            )
-
-        case .zStack:
-            layerType = .group
-            extras.append(
-                .layerInputSet(VPLLayerInputSet(id: id,
-                                                input: .orientation,
-                                                value: .orientation(.none)))
-            )
-
-        case .roundedRectangle:
-            layerType = .rectangle
-            if let arg = constructorArguments.first(where: { $0.label == .cornerRadius }),
-               let radius = Double(arg.value) {
-                extras.append(
-                    .layerInputSet(VPLLayerInputSet(id: id,
-                                                    input: .cornerRadius,
-                                                    value: .number(radius)))
-                )
-            }
-
-        default:
-            layerType = .hitArea   // safe fallback
-        }
-
-        // ── Generic constructor‑argument handling (literals & edges) ─────────
-        for arg in constructorArguments {
-
-            // Skip the specialised RoundedRectangle .cornerRadius
-            if self.name == .roundedRectangle && arg.label == .cornerRadius { continue }
-
-            guard let port = arg.deriveLayerInputPort(layerType),
-                  let portValue = arg.derivePortValue(layerType) else { continue }
-
-            switch arg.syntaxKind {
-            case .literal:
-                extras.append(
-                    .layerInputSet(VPLLayerInputSet(id: id,
-                                                    input: port,
-                                                    value: portValue))
-                )
-            case .variable, .expression:
-                extras.append(.incomingEdge(VPLIncomingEdge(name: port)))
-            }
-        }
-
-        // ── Generic modifier handling ────────────────────────────────────────
-        for modifier in modifiers {
-
-            guard let port = modifier.name.deriveLayerInputPort(layerType) else { continue }
-
-            // Start with default value for that port
-            var portValue = port.getDefaultValue(for: layerType)
-
-            if modifier.arguments.count == 1, let arg = modifier.arguments.first {
-
-                var raw = arg.value
-                if let c = Color.fromSystemName(raw) { raw = c.asHexDisplay }
-                let input = PortValue.string(.init(raw))
-
-                let coerced = [input].coerce(to: portValue, currentGraphTime: .zero)
-                if let first = coerced.first { portValue = first }
-
-            } else {
-                for (idx, arg) in modifier.arguments.enumerated() {
-                    portValue = portValue.parseInputEdit(
-                        fieldValue: .string(.init(arg.value)),
-                        fieldIndex: idx
-                    )
-                }
-            }
-
-            extras.append(
-                .layerInputSet(VPLLayerInputSet(id: id,
-                                                input: port,
-                                                value: portValue))
-            )
-        }
-
-        // Final bare layer (children added later)
-        return (VPLLayer(id: id, name: layerType, children: []), extras)
-    }
-    
     /// Recursive conversion to **flattened** `[VPLLayerConcept]`
-    func recursivelyDeriveActions() -> VPLLayerConcepts {
+    func recursivelyDeriveActions() -> VPLLayerConcepts? {
         
         // 1. Map this node
-        var (layer, concepts) = deriveLayer()
-        var childLayers: [VPLLayer] = []
-        
-        // 2. Recurse into children
-        for child in children {
-            let childConcepts = child.recursivelyDeriveActions()
+        if var (layer, concepts) = deriveLayer() {
+            var childLayers: [VPLLayer] = []
             
-            // First concept for every child must be its `.layer`
-            if case let .layer(childLayer) = childConcepts[0] {
-                childLayers.append(childLayer)
+            // 2. Recurse into children
+            for child in children {
+                if let childConcepts = child.recursivelyDeriveActions() {
+                    // First concept for every child must be its `.layer`
+                    if case let .layer(childLayer) = childConcepts[0] {
+                        childLayers.append(childLayer)
+                    }
+                    concepts.append(contentsOf: childConcepts) // depth-first
+                }
             }
-            concepts.append(contentsOf: childConcepts) // depth-first
+            
+            let layerWithChildren = VPLLayer(id: layer.id,
+                                             name: layer.name,
+                                             children: childLayers)
+            
+            // 3. Prepend *this* fully-assembled layer concept
+            concepts.insert(.layer(layerWithChildren), at: 0)
+            
+            return concepts
+        } else {
+            return nil
         }
         
-        let layerWithChildren = VPLLayer(id: layer.id,
-                                         name: layer.name,
-                                         children: childLayers)
-        
-        // 3. Prepend *this* fully-assembled layer concept
-        concepts.insert(.layer(layerWithChildren), at: 0)
-        
-        return concepts
     }
 }
 
 extension SyntaxView {
     
-    func deriveStitchActions() -> VPLLayerConceptOrderedSet {
-        let concepts = recursivelyDeriveActions()
-        return VPLLayerConceptOrderedSet(concepts)
+    func deriveStitchActions() -> VPLLayerConceptOrderedSet? {
+        if let concepts = recursivelyDeriveActions() {
+            return VPLLayerConceptOrderedSet(concepts)
+        } else {
+            return nil
+        }
     }
 }
